Rename course form components and add course page to create course page
